feat(cli): Improve tool call result rendering
Enhanced tool call result handling with configurable inline display
options. Tool call results can be either fully rendered inline,
truncated to a configurable number of lines, or not rendered at all. In
all three cases, the result will a) be fully returned to the assistant,
and b) be persisted to a file for further inspection. A link to the
local file is rendered inline.
